JPERF-904: Apply axion-release plugin's instructions to CI config
The release process of `report` module stopped working likely because we don't have up to date token in repository's secrets storage.
So far we needed to update this token manually which we don't want to do, as it requires knowledge maintenance and is doomed to fail at some point.

We are using the token in `axion-release` gradle plugin which is dependency of `gradle-release` gradle plugin developed by us.
It is used to create git tags of releases.

With `axion-release` it's possible to reuse github automated mechanisms to authenticate git client operations.
The way to do that is documented in https://axion-release-plugin.readthedocs.io/en/latest/configuration/ci_servers/#github-actions

In order to make the `github.token` work for git client mutable operations like pushing tags we need permission `contents: write` for the CI job.

I decided to not grant this permission to the rest of the CI (tests), so I extracted the `release` job.
This way tests kept the default permissions defined on the repository level.