JPERF-1059: Handle multiple JQL `ActionType`s
`JqlReport` is more open-ended, because it doesn't expose its
constructor. It's possible to change the way it writes CSV,
the stats it calculates, etc. because the Builder allows new parameters.

Which params should be passed via constructors and which via methods?
I think that constructors should hold the caller preferences.
The content of `ActionMetric`s is not a preference, it comes from
another part of the system, from measurements.
The `target: Path` is also calculated by a convention of `FullReport`,
not a user preference.

But which `List<ActionType>` cannot be calculated by the system
automatically, because Java has generic type erasure.
The user has to provide candidates for `ActionType<SearchJqlObservation>`.
It has to be a user preference.