JPERF-1160: Test noise case for typical perf judge
I assume the alpha/beta test results represent an A vs A pair.
They need to be considered neither regression nor improvement.
They should be considered noise.

The `LatencyImpact.conclusive` was difficult to interpret in the case of
noise. Could we _conclude_ the regression _doesn't_ exist?

Now we explicitly talk about [signal and noise], which is a well known
paradigm in engineering.

[signal and noise]: https://en.wikipedia.org/wiki/Signal-to-noise_ratio

TDD: red