JPERF-1160: Stop expecting confidence in noise
When a judge doesn't detect any impact: no regression, no improvement,
then we were calling it "noise". But it can be a real result, e.g. when
you test 9.8.0 vs 9.8.0 then "no impact" is actually expected.
We have a math model for confidence of:
* [regression] vs [no regression (improvement or no impact)]
* [improvement] vs [no improvement (regression or no impact)]
But we don't have a clear model for "no impact" itself.
So for now we refuse to caluclate confidence for "no impact".

TDD: red