JPERF-1160: Rename "noise" to "irrelevant"
"Noise" implies that the test is wrong and needs to be fixed.
When you test two systems with the same performance, then it's not
noise, but lack of performance impact.

First, we renamed "noise" to "no impact", but `LatencyImpact.impact`
and `LatencyImpact.noImpact` looked weird and was hard to interpret.
So we settled on `LatencyImpact.relevant` + `LatencyImpact.irrelevant`.

In the latency markdown table, we still print "NO IMPACT",
because "IRRELEVANT" was too similar to "IMPROVEMENT" visually.

TDD: refactor