JPERF-1160: Admit to possible inconclusive results
It's possible that nothing dominates: not regressions, not improvements,
not "irrelevants". Then we're truly inconclusive.
Such a situation could happen even before 3-valued logic.
Back when we either had a binary regression or non-regression",
it would be legit to classify that as both a 0% confidence regression
and a 0% confidence non-regression at the same time.

Technically, we could pick the most frequent class, e.g. for:
* 5 regressions
* 4 improvements
* 3 irrelevants
We could say it's a 5 vs 7 = 43% confidence regression.
But if picking a non-dominated class is ok, then we could ask
if we assume it's an improvement, how confident are we?
It would be 4 improvements vs 8 non-improvements,
but the current math would say it's 75,2% confident,
which is an undeserved confidence boost.

Picking a non-dominated case doesn't solve classification ambiguity.
E.g. when nothing dominates and there's a tie: 4-4-3 or 4-4-4.

It's hard to say what should happen, so for now
let's say that non-dominated cases are inconclusive.

It's greener, but still red.

TDD: red