JPERF-1409 Give up on matching profilers with nodes
The `EdibleResult.nodeDistribution` DOES NOT map
with `ActionMetric.virtualUser`.
It only counts VUs, without listing their ids.
Moreover, it assumes sticky sessions, ie. a given VU interacts only with
a single node. It's currently true for cookie/browser traffic in DC.
But it's not relevant for stateless REST traffic now
or for human traffic in the future.

So for now, we don't filter by node id and the investigating dev will
have to inspect all profiles for all nodes.
The thread id and timeslot filtering reduces the noise significantly,
but not maximally (yet).